compile warnings

+++ tsk/fs/hfs_dent.c
@@ -265,7 +265,7 @@ hfs_dir_open_meta_cb(HFS_INFO * hfs, int8_t level_type,
 
             // Make sure there is enough space in cur_key for the name 
             // (name is unicode so each characters is two bytes; 6 bytes of non-name characters)
-            if ((uint32_t)(tsk_getu16(hfs->fs_info.endian, cur_key->name.length)) * 2 > tsk_getu16(hfs->fs_info.endian, cur_key->key_len) - 6) {
+            if ((uint32_t)(tsk_getu16(hfs->fs_info.endian, cur_key->name.length)) * (unsigned int)2 > tsk_getu16(hfs->fs_info.endian, cur_key->key_len) - 6) {
                 error_returned
                 ("hfs_dir_open_meta_cb: name length is too long");
                 return HFS_BTREE_CB_ERR;
@@ -323,7 +323,7 @@ hfs_dir_open_meta_cb(HFS_INFO * hfs, int8_t level_type,
 
             // Make sure there is enough space in cur_key for the name 
             // (name is unicode so each characters is two bytes; 6 bytes of non-name characters)
-            if ((uint32_t)(tsk_getu16(hfs->fs_info.endian, cur_key->name.length)) * 2 > tsk_getu16(hfs->fs_info.endian, cur_key->key_len) - 6) {
+            if ((uint32_t)(tsk_getu16(hfs->fs_info.endian, cur_key->name.length)) * (unsigned int)2 > tsk_getu16(hfs->fs_info.endian, cur_key->key_len) - 6) {
                 error_returned
                 ("hfs_dir_open_meta_cb: name length is too long");
                 return HFS_BTREE_CB_ERR;
